+++ Interfaces/IDamageable.cs
@@ -5,16 +5,16 @@ public interface IDamageable
     /// <summary>
     /// Current health of the damageable.
     /// </summary>
-    protected float Health { get; set; }
+    protected double Health { get; set; }
     
     /// <summary>
     /// Maximum health of the damageable.
     /// </summary>
-    protected float MaxHealth { get; }
+    protected double MaxHealth { get; }
     
     /// <summary>
     /// Take damage.
     /// </summary>
     /// <param name="damage">The damage to take.</param>
-    public void TakeDamage(float damage);
+    public void TakeDamage(double damage);
 }
